+++ app/services/AddressLookupService.scala
@@ -20,15 +20,21 @@ import com.kenshoo.play.metrics.Metrics
 import config.ConfigDecorator
 import com.google.inject.{Inject, Singleton}
 import metrics._
-import models.addresslookup.{AddressLookupErrorResponse, AddressLookupResponse, AddressLookupSuccessResponse, AddressLookupUnexpectedResponse, RecordSet}
+import models.addresslookup.RecordSet
 import play.api.Mode.Mode
 import play.api.{Configuration, Environment, Logger}
 import services.http.SimpleHttp
-import uk.gov.hmrc.http.HeaderCarrier
+import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 import util._
 
 import scala.concurrent.Future
+
+sealed trait AddressLookupResponse
+
+final case class AddressLookupSuccessResponse(addressList: RecordSet) extends AddressLookupResponse
+final case class AddressLookupUnexpectedResponse(r: HttpResponse) extends AddressLookupResponse
+final case class AddressLookupErrorResponse(cause: Exception) extends AddressLookupResponse
 
 @Singleton
 class AddressLookupService @Inject()(
