+++ app/views/cards/personaldetails/mainAddress.scala.html
@@ -19,7 +19,7 @@
 @(personDetails: models.PersonDetails, taxCreditsEnabled: Boolean, hasCorrespondenceAddress: Boolean, isLocked: Boolean, countries: List[Country])(implicit messages: play.api.i18n.Messages)
 
     @updateAddressLink() = @{
-        if(taxCreditsEnabled) controllers.address.routes.TaxCreditsChoiceController.onPageLoad() else controllers.address.routes.ResidencyChoiceController.onPageLoad()
+        if(taxCreditsEnabled) controllers.routes.AddressController.taxCreditsChoice else controllers.routes.AddressController.residencyChoice
     }
 
     @cardContent = {
@@ -54,7 +54,7 @@
                 </li>
                 @if(!hasCorrespondenceAddress) {
                     <li>
-                        <a class="ga-track-anchor-click" href="@controllers.address.routes.InternationalAddressChoiceController.onPageLoad(PostalAddrType)" data-ga-event-category="link - click" data-ga-event-action="Main address" data-ga-event-label="Change where we send your letters">@messages("label.change_where_we_send_your_letters")</a>
+                        <a class="ga-track-anchor-click" href="@controllers.routes.AddressController.internationalAddressChoice(PostalAddrType)" data-ga-event-category="link - click" data-ga-event-action="Main address" data-ga-event-label="Change where we send your letters">@messages("label.change_where_we_send_your_letters")</a>
                     </li>
                 }
             </ul>
