+++ app/views/personaldetails/cannotUpdateAddress.scala.html
@@ -34,7 +34,7 @@ partialRetriever: LocalPartialRetriever, templateRenderer: TemplateRenderer, mes
 
       <p class="lede margin-bottom">@messages("label.the_date_you_entered_is_earlier_than_a_date_previously_held_")</p>
 
-      <p class="margin-top">@messages("label.if_you_entered_proposedStartDate_by_mistake_text_before_link_1") @proposedStartDate @Messages("label.if_you_entered_proposedStartDate_by_mistake_text_before_link_2") <a href="@controllers.address.routes.StartDateController.onPageLoad(typ).url">@messages("label.if_you_entered_proposedStartDate_by_mistake_link_text")</a>.</p>
+      <p class="margin-top">@messages("label.if_you_entered_proposedStartDate_by_mistake_text_before_link_1") @proposedStartDate @Messages("label.if_you_entered_proposedStartDate_by_mistake_text_before_link_2") <a href="@controllers.routes.AddressController.enterStartDate(typ).url">@messages("label.if_you_entered_proposedStartDate_by_mistake_link_text")</a>.</p>
 
       <p>@messages("label.if_you_entered_the_correct_date_contact_HMRC_text_before_link") <a href="@configDecorator.changePersonalDetailsUrl" target="_blank" rel="noopener noreferrer">@messages("label.if_you_entered_the_correct_date_contact_HMRC_link_text") @messages("label.opens_in_a_new_window")</a> @messages("label.if_you_entered_the_correct_date_contact_HMRC_text_after_link").</p>
