+++ app/views/personaldetails/reviewChanges.scala.html
@@ -45,7 +45,7 @@ partialRetriever: LocalPartialRetriever, templateRenderer: TemplateRenderer, mes
             @messages("label.do_you_live_in_the_uk")
           </dt>
           <dd class="cya-answer">@messages(doYouLiveInTheUK)</dd>
-          <dd class="cya-change"><a href="@controllers.address.routes.InternationalAddressChoiceController.onPageLoad(typ)">@messages("label.change")</a></dd>
+          <dd class="cya-change"><a href="@controllers.routes.AddressController.internationalAddressChoice(typ)">@messages("label.change")</a></dd>
         </div>
         <div>
           <dt class="cya-question">
@@ -69,12 +69,12 @@ partialRetriever: LocalPartialRetriever, templateRenderer: TemplateRenderer, mes
 
             <dd class="cya-answer">@Html(addressDto.toList.map(item => s"<div>$item</div>").mkString)</dd>
 
-            <dd class="cya-change"><a href="@controllers.address.routes.UpdateAddressController.onPageLoad(typ)">@messages("label.change_address")</a></dd>
+            <dd class="cya-change"><a href="@controllers.routes.AddressController.showUpdateAddressForm(typ)">@messages("label.change_address")</a></dd>
             } else {
 
             <dd class="cya-answer">@Html(addressDto.toListWithCountry.map(item => s"<div>$item</div>").mkString)</dd>
 
-            <dd class="cya-change"><a href="@controllers.address.routes.UpdateInternationalAddressController.onPageLoad(typ)">@messages("label.change")</a></dd>
+            <dd class="cya-change"><a href="@controllers.routes.AddressController.showUpdateInternationalAddressForm(typ)">@messages("label.change")</a></dd>
             }
 
         </div>
@@ -85,13 +85,13 @@ partialRetriever: LocalPartialRetriever, templateRenderer: TemplateRenderer, mes
                   @typ.ifIs(primary = messages("label.when_this_became_your_main_home"), sole = messages("label.when_you_started_living_here"), postal = "")
                 </dt>
                 <dd class="cya-answer">@LanguageHelper.langUtils.Dates.formatDate(startDateDto.startDate)</dd>
-                <dd class="cya-change"><a href="@controllers.address.routes.StartDateController.onPageLoad(typ)">@messages("label.change")</a></dd>
+                <dd class="cya-change"><a href="@controllers.routes.AddressController.enterStartDate(typ)">@messages("label.change")</a></dd>
               }
             </div>
           }
       </dl>
 
-      @form(controllers.address.routes.AddressSubmissionController.onSubmit(typ)) {
+      @form(controllers.routes.AddressController.submitChanges(typ)) {
         <button type="submit">@messages("label.confirm_and_save")</button>
       }
     </div>
