+++ test/controllers/ApplicationControllerSpec.scala
@@ -17,7 +17,6 @@
 package controllers
 
 import config.ConfigDecorator
-import controllers.address.TaxCreditsChoiceController
 import controllers.auth.requests.UserRequest
 import controllers.auth.{AuthAction, AuthJourney, SelfAssessmentStatusAction}
 import models._
@@ -54,8 +53,8 @@ class ApplicationControllerSpec extends BaseSpec with CurrentTaxYear with Mockit
   val mockSelfAssessmentStatusAction = mock[SelfAssessmentStatusAction]
   val mockAuthJourney = mock[AuthJourney]
   val mockInterstitialController = mock[InterstitialController]
+  val mockAddressController = mock[AddressController]
   val mockHomeController = mock[HomeController]
-  val mockTaxCreditsChoiceController = mock[TaxCreditsChoiceController]
 
   override implicit lazy val app: Application = localGuiceApplicationBuilder()
     .overrides(
@@ -64,8 +63,8 @@ class ApplicationControllerSpec extends BaseSpec with CurrentTaxYear with Mockit
       bind[SelfAssessmentStatusAction].toInstance(mockSelfAssessmentStatusAction),
       bind[AuthJourney].toInstance(mockAuthJourney),
       bind[InterstitialController].toInstance(mockInterstitialController),
-      bind[HomeController].toInstance(mockHomeController),
-      bind[TaxCreditsChoiceController].toInstance(mockTaxCreditsChoiceController)
+      bind[AddressController].toInstance(mockAddressController),
+      bind[HomeController].toInstance(mockHomeController)
     )
     .build()
